pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[project]
name = "eldritch-tactics"
version = "0.1.0"
description = "A turn-based tactical game inspired by X-COM, featuring squads of flawed investigators fighting against Lovecraftian horrors"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["game", "tactics", "lovecraft", "pygame", "turn-based", "strategy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Games/Entertainment :: Turn Based Strategy",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pygame-ce>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
eldritch-tactics = "main:main"

[project.urls]
Homepage = "https://github.com/yourusername/eldritch-tactics"
Repository = "https://github.com/yourusername/eldritch-tactics"
Issues = "https://github.com/yourusername/eldritch-tactics/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=term-missing",
]

[tool.uv]
# UV-specific configuration
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
